object_detector_app/requirements.txt
pillow
numpy
Flask
ultralytics
